poo ex2 feito - java e python

+++ poo/java/ex2-funcionario-java/Vendedor.java
@@ -1,7 +1,7 @@
 
 public class Vendedor extends Funcionario  {
-    long comissao;
-    public Vendedor(String nome, String cpf, double salarioBase, long comissao_valor)
+    double comissao;
+    public Vendedor(String nome, String cpf, double salarioBase, double comissao_valor)
     {
         setInformacoes(nome, cpf, salarioBase);
         comissao = comissao_valor;
